+++ examples/pipe_example.py
@@ -1,0 +1,63 @@
+"""
+##### A very simply data pipeline setup using `kucoincli.pipe` module #####
+
+Database created will look like this ...... 
+
++----------+-------+-------+-------+-------+-------------+ 
+|   time   |  open | close |  high |  low  |   volume    |  
++----------+-------+-------+-------+-------+-------------+ 
+|2022-01-01|1001.51|1002.21|1008.32| 999.43|8.14505485512|
+|2022-01-02| 999.49|1000.80|1004.89| 995.32|9.15848158419|
+...
+...
+...
+"""
+
+# Import our modules
+from kucoincli.pipe import pipeline
+from sqlalchemy import create_engine
+import logging
+
+# Create our sqlite database engine with sqlalchemy.
+# This engine will generate a new database for us if 
+# none exists prior.
+# Be aware that it is highly recommended to use a true
+# SQL database such as Postgre.
+engine = create_engine("sqlite:///example.db")
+
+# Add a logger to see pick up some additional output info
+# To retrieve timeout messages set logging level to DEBUG
+fmt = "%(asctime)s [%(levelname)s] %(module)s :: %(message)s"
+logging.basicConfig(level=logging.INFO, format=fmt)
+logging.getLogger(__name__)
+
+# Setup our constants for the pipeline.
+# There are many more adjustments we can
+# make to the pipe, but we want to keep this 
+# example as simple as possible.
+TICKER = "BT-USDT" # Note we could specify multiple tickers as a list
+START = "2022-04-01" # Note we could specify this as a datetime object
+END = "2022-05-01"
+INTERVAL = "1min"
+
+# Now let's ppen up our pipeline ...
+# Note that we are not specifying a schema as SQLite will not allow it
+pipeline(
+    tickers=TICKER,     # Tickers to query OHLCV data for
+    engine=engine,      # Engine to run our database
+    interval=INTERVAL,  # Interval at which to obtain OHLCV
+    start=START,        # Earliest date to query from
+    end=END,            # Latest date to query from
+)
+
+# Viola, we have generated a database. 
+
+# This pipeline will obtain 30 days of minutely data and store it in
+# an SQLite .db file in our examples folder. It will conviently
+# demonstrate its progress with a progress bar and let the user know 
+# when it has finished retrieving and filing the data.
+
+## Thats the entire pipeline! As easy as that we have created a permanent ##
+## SQLite database to draw from for future research. See the SQLalchemy   ##  
+## documentation (https://docs.sqlalchemy.org/en/14/) for details on how  ##
+## to interact with our new database.                                     ##