+++ docs/changelog.rst
@@ -1,2 +1,8 @@
 Changelog
 =========
+
+1.1.0
+-----
+* Completely reworked `kucoincli.pipe`
+    * Made `schema` optional
+    * Added functionality